refactor/#473: navigate 관련 코드 훅 분리
