+++ frontend/src/pages/ReviewStarRating/ReviewStarRating.tsx
@@ -1,27 +1,14 @@
-import { useLocation, useNavigate } from 'react-router-dom';
 import { styled } from 'styled-components';
 
 import PageHeader from '@/components/@common/PageHeader/PageHeader';
 import StarRatingInput from '@/components/@common/StarRating/StarRatingInput/StarRatingInput';
 import Template from '@/components/@common/Template';
 import { useReviewStarRating } from '@/hooks/review/useReviewStarRating';
-import { routerPath } from '@/router/routes';
 
 const ReviewStarRating = () => {
-  const navigate = useNavigate();
-  const location = useLocation();
-  const { petFoodId, foodData, isEditMode, reviewId, rating, setRating } =
-    useReviewStarRating(location);
+  const { foodData, rating, setRating, onClickStar, goBack } = useReviewStarRating();
 
   const onMouseDownStar = (selectedRating: number) => setRating(selectedRating);
-
-  const onClickStar = (selectedRating: number) => {
-    navigate(routerPath.reviewAddition({ petFoodId }), {
-      state: { selectedRating, isEditMode, reviewId },
-    });
-  };
-
-  const goBack = () => navigate(routerPath.back);
 
   if (!foodData) throw new Error('죄송합니다, 해당 식품 정보를 찾을 수 없습니다.');
